refactor admin & use common table
