+++ client/src/components/admin-panel/AdminPanel.js
@@ -2,16 +2,16 @@ import React, { Component } from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import Filters from './Filters';
-import FiltersDetail from './FiltersDetail';
-import Products from './Products';
-import ProductsDetail from './ProductsDetail';
-import Categories from './Categories';
-import CategoriesDetail from './CategoriesDetail';
-import Footer from './Footer';
-import FooterDetail from './FooterDetail';
-import Slider from './Slider';
-import SliderDetail from './SliderDetail';
+import Filters from './Filters/Filters';
+import FiltersDetail from './Filters/FiltersDetail';
+import Products from './Products/Products';
+import ProductsDetail from './Products/ProductsDetail';
+import Categories from './Categories/Categories';
+import CategoriesDetail from './Categories/CategoriesDetail';
+import Footer from './Footer/Footer';
+import FooterDetail from './Footer/FooterDetail';
+import Slider from './Slider/Slider';
+import SliderDetail from './Slider/SliderDetail';
 import PayMethod from './PayMethod/PayMethod';
 import NotFound from '../page/NotFound';
 import ShippingMethod from './ShippingMethod/ShippingMethod';
