+++ client/src/components/admin-panel/Categories/Categories.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { Redirect } from 'react-router';
 
-import BtnCreateAdmin from './../common/admin-panel/BtnCreateAdmin';
+import BtnCreateAdmin from '../../common/admin-panel/BtnCreateAdmin';
 
-import SnackBars from '../common/admin-panel/SnackBars';
-import Preloader from '../common/admin-panel/Preloader';
+import SnackBars from '../../common/admin-panel/SnackBars';
+import Preloader from '../../common/admin-panel/Preloader';
 
-import AdminCategoriesAPI from '../../services/AdminCategoriesAPI';
+import AdminCategoriesAPI from '../../../services/AdminCategoriesAPI';
 
 import MaterialTable from 'material-table';
 
@@ -14,7 +14,7 @@ import Switch from '@material-ui/core/Switch';
 
 import DeleteOutline from '@material-ui/icons/DeleteOutline';
 
-import { tableIcons } from './TableIcons';
+import { tableIcons } from '../../common/admin-panel/TableIcons';
 
 export default class Categories extends Component {
   state = {
