+++ client/src/components/admin-panel/Categories/CategoriesDetail.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import AdminFiltersAPI from '../../services/AdminFiltersAPI';
-import AdminCategoriesAPI from '../../services/AdminCategoriesAPI';
+import AdminFiltersAPI from '../../../services/AdminFiltersAPI';
+import AdminCategoriesAPI from '../../../services/AdminCategoriesAPI';
 
 import CategoriesDetailForm from './CategoriesDetailForm.js';
 
-import SnackBars from '../common/admin-panel/SnackBars';
-import Preloader from '../common/admin-panel/Preloader';
+import SnackBars from '../../common/admin-panel/SnackBars';
+import Preloader from '../../common/admin-panel/Preloader';
 
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
