+++ client/src/components/admin-panel/Categories/CategoriesDetailForm.js
