+++ client/src/components/admin-panel/ConfigSystem/ConfigSystem.js
@@ -5,7 +5,7 @@ import Dialog from '@material-ui/core/Dialog';
 import DialogContent from '@material-ui/core/DialogContent';
 import ConfigSystemDetail from './ConfigSystemDetail/ConfigSystemDetail';
 import Switch from '@material-ui/core/Switch';
-import { tableIcons } from '../TableIcons';
+import { tableIcons } from '../../common/admin-panel/TableIcons';
 import MaterialTable from 'material-table';
 import ConfigurationAPI from '../../../services/ConfigurationAPI';
 import { Button, Typography } from '@material-ui/core';
