+++ client/src/components/admin-panel/Filters/FiltersDetail.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import AdminFiltersAPI from '../../services/AdminFiltersAPI';
+import AdminFiltersAPI from '../../../services/AdminFiltersAPI';
 
 import FiltersDetailForm from './FiltersDetailForm';
 
-import SnackBars from '../common/admin-panel/SnackBars';
-import Preloader from '../common/admin-panel/Preloader';
+import SnackBars from '../../common/admin-panel/SnackBars';
+import Preloader from '../../common/admin-panel/Preloader';
 
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
