+++ client/src/components/admin-panel/Filters/FiltersDetailForm.js
