+++ client/src/components/admin-panel/Footer/Footer.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { Redirect } from 'react-router';
 
-import BtnCreateAdmin from './../common/admin-panel/BtnCreateAdmin';
+import BtnCreateAdmin from '../../common/admin-panel/BtnCreateAdmin';
 
-import SnackBars from '../common/admin-panel/SnackBars';
-import Preloader from '../common/admin-panel/Preloader';
+import SnackBars from '../../common/admin-panel/SnackBars';
+import Preloader from '../../common/admin-panel/Preloader';
 
-import AdminFiltersAPI from '../../services/AdminFiltersAPI';
+import AdminFooterAPI from '../../../services/AdminFooterAPI';
 
 import MaterialTable from 'material-table';
 
@@ -14,13 +14,14 @@ import Switch from '@material-ui/core/Switch';
 
 import DeleteOutline from '@material-ui/icons/DeleteOutline';
 
-import { tableIcons } from './TableIcons';
+import { tableIcons } from '../../common/admin-panel/TableIcons';
 
-export default class Filters extends Component {
+export default class Footer extends Component {
   state = {
     columns: [
-      { title: 'Filter', field: 'title' },
-      { title: 'Service name', field: 'serviceName' },
+      { title: 'Group link', field: 'title' },
+      { title: 'Link', field: 'description' },
+      { title: 'Url', field: 'url' },
       {
         title: 'Enabled',
         field: 'enabled',
@@ -57,22 +58,22 @@ export default class Filters extends Component {
 
       delData.forEach(async item => {
         if (item.parentId) {
-          await AdminFiltersAPI.deleteSubFilters(item.id);
+          await AdminFooterAPI.deleteFooterLink(item.parentId, { _id: item.id });
 
           this.setIsLoading(false);
 
           this.setState({
             sendDataStatus: 'success',
-            sendDataMessage: `${item.title} sub filter has been remove!`
+            sendDataMessage: `${item.title} group link has been remove!`
           });
         } else {
-          await AdminFiltersAPI.deleteFilters(item.id);
+          await AdminFooterAPI.deleteFooterGroupLinks(item.id);
 
           this.setIsLoading(false);
 
           this.setState({
             sendDataStatus: 'success',
-            sendDataMessage: `${item.title} filter has been remove!`
+            sendDataMessage: `${item.title} link has been remove!`
           });
 
           this.setState(prevState => {
@@ -92,31 +93,31 @@ export default class Filters extends Component {
   };
 
   onRowClick = (evt, selectedRow) => {
-    this.setState({ clickId: selectedRow.id });
+    this.setState({ clickId: selectedRow.parentId ? selectedRow.parentId : selectedRow.id });
   };
 
   onRefreshData = async () => {
     try {
       this.setIsLoading(true);
 
-      const { data } = await AdminFiltersAPI.getFilters();
+      const { data } = await AdminFooterAPI.getFooter();
 
       const preViewRes = [];
 
-      data.forEach(item => {
+      data.forEach(group => {
         preViewRes.push({
-          id: item._id,
-          title: item.type,
-          serviceName: item.serviceName,
-          enabled: item.enabled
+          id: group._id,
+          title: group.title,
+          enabled: group.enabled
         });
 
-        item._idSubFilters.forEach(sub => {
+        group.links.forEach(link => {
           preViewRes.push({
-            id: sub._id,
-            title: sub.name,
-            parentId: item._id,
-            enabled: sub.enabled
+            id: link._id,
+            description: link.description,
+            url: link.url,
+            parentId: group._id,
+            enabled: link.enabled
           });
         });
       });
@@ -160,7 +161,7 @@ export default class Filters extends Component {
       <>
         <MaterialTable
           icons={tableIcons}
-          title="Filters"
+          title="Footer"
           columns={columns}
           data={data}
           parentChildData={(row, rows) => rows.find(a => a.id === row.parentId)}
@@ -191,7 +192,7 @@ export default class Filters extends Component {
           ]}
           onRowClick={this.onRowClick}
         />
-        <BtnCreateAdmin to="/admin-panel/filters/new" />
+        <BtnCreateAdmin to="/admin-panel/footer/new" />
 
         <SnackBars
           handleClose={this.handleCloseSnackBars}
@@ -202,9 +203,7 @@ export default class Filters extends Component {
 
         <Preloader open={isLoading} />
 
-        {this.state.clickId ? (
-          <Redirect to={`/admin-panel/filters/${clickId}`} push={true} />
-        ) : null}
+        {this.state.clickId ? <Redirect to={`/admin-panel/Footer/${clickId}`} push={true} /> : null}
       </>
     );
   }
