+++ client/src/components/admin-panel/Footer/FooterDetailForm.js
