+++ client/src/components/admin-panel/Products/ProductsDetail.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
-import AdminCategoriesAPI from '../../services/AdminCategoriesAPI';
-import AdminProductsAPI from '../../services/AdminProductsAPI';
+import AdminCategoriesAPI from '../../../services/AdminCategoriesAPI';
+import AdminProductsAPI from '../../../services/AdminProductsAPI';
 
-import SnackBars from '../common/admin-panel/SnackBars';
-import Preloader from '../common/admin-panel/Preloader';
+import SnackBars from '../../common/admin-panel/SnackBars';
+import Preloader from '../../common/admin-panel/Preloader';
 
 import ProductsDetailBasicInfo from './ProductsDetailBasicInfo';
 import ProductsDetailMainImages from './ProductsDetailMainImages';
