+++ client/src/components/admin-panel/Products/ProductsDetailBasicInfo.js
