+++ client/src/components/admin-panel/Products/ProductsDetailMainImages.js
