+++ client/src/components/admin-panel/Slider/SliderDetailForm.js
